requirements.txt
webdriver-manager
selenium
pytest==8.4.2
pytest-repeat==0.9.4
pytest-rerunfailures==16.0.1
pytest-xdist == 3.8.0
allure-pytest == 2.15.0

# dev
black==25.1.0
flake8==7.3.0
isort==6.0.1
pre-commit==4.3.0
